added navigation stack
