+++ screens/lmao.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+ import {
+   SafeAreaView,
+   StyleSheet,
+   Text,
+ } from 'react-native';
+ 
+ 
+ const LmaoScreen = () => {
+ 
+   return (
+     <SafeAreaView>
+      <Text>
+          LMAO
+      </Text>
+     </SafeAreaView>
+   );
+ };
+ 
+ const styles = StyleSheet.create({
+ });
+ 
+ export default LmaoScreen;
+
